+++ custom_components/xtend_tuya/const.py
@@ -106,6 +106,7 @@ PLATFORMS = [
     Platform.ALARM_CONTROL_PANEL,
     Platform.BINARY_SENSOR,
     Platform.BUTTON,
+    Platform.CALENDAR,
     Platform.CAMERA,
     Platform.CLIMATE,
     Platform.COVER,